Cleanup and info/error messages

+++ src/features/blogs/Blog.js
@@ -2,7 +2,10 @@ import React, { useState } from 'react'
 import PropTypes from 'prop-types'
 import { useDispatch, useSelector } from 'react-redux'
 import { addLikeToBlog, removeBlog } from './blogsSlice'
-import { sendErrorMessage } from '../notification/notificationSlice'
+import {
+  sendErrorMessage,
+  sendInfoMessage,
+} from '../notification/notificationSlice'
 import { unwrapResult } from '@reduxjs/toolkit'
 import { selectUser } from '../users/usersSlice'
 
@@ -24,8 +27,20 @@ const Blog = ({ blog }) => {
       })
   }
 
-  const handleRemoveButton = () => {
-    dispatch(removeBlog(blog))
+  const handleRemoveButton = async () => {
+    const confirmation = window.confirm(
+      `Are you sure you want to delete the blog ${blog.title} by ${blog.author} ?`
+    )
+    if (confirmation) {
+      try {
+        const resultRemoveBlog = await dispatch(removeBlog(blog))
+        unwrapResult(resultRemoveBlog)
+        const { meta } = resultRemoveBlog
+        dispatch(sendInfoMessage(`The blog ${meta.title} has been removed`))
+      } catch (error) {
+        dispatch(sendErrorMessage(error.message))
+      }
+    }
   }
 
   const removeButton = () => {
